+++ samples/EasyOBJDTuner.java
@@ -15,7 +15,7 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.easyobd.EasyOBJD;
 import org.firstinspires.ftc.easyobd.EasyOBJDPipeline;
-import org.firstinspires.ftc.easyobd.OverlayMode;
+import org.firstinspires.ftc.easyobd.EasyOBJD.OverlayMode;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;